+++ R/utils,conditions.R
@@ -3,7 +3,7 @@ stopf <- function(fmt, ..., call. = TRUE, domain = NULL) {  #nolint
   msg <- .makeMessage(msg, domain = domain)
   if (is.call(call.)) {
     call <- call.
-  } else if (isTRUE(call)) {
+  } else if (isTRUE(call.)) {
     call <- sys.call(which = -1L)
   } else {
     call <- NULL
@@ -21,7 +21,7 @@ warnf <- function(fmt, ..., call. = TRUE, immediate. = FALSE, domain = NULL) {  
     msg <- .makeMessage(msg, domain = domain)
     if (is.call(call.)) {
       call <- call.
-    } else if (isTRUE(call)) {
+    } else if (isTRUE(call.)) {
       call <- sys.call(which = -1L)
     } else {
       call <- NULL
@@ -29,9 +29,4 @@ warnf <- function(fmt, ..., call. = TRUE, immediate. = FALSE, domain = NULL) {  
     cond <- simpleWarning(msg, call = call)
     warning(cond)
   }
-}
-
-msgf <- function(fmt, ..., appendLF = FALSE, domain = NULL) {  #nolint
-  msg <- gettextf(fmt, ...)
-  message(msg, appendLF = appendLF, domain = domain)
 }